// extensions/wecom/src/capability/calendar/types.ts
export interface CalendarSharesEntry {
  userid: string;
  permission?: 1 | 3; // 1: 可查看，3: 仅查看闲忙状态
}

export interface CalendarPublicRange {
  userids?: string[]; // 公开的成员列表，最多 1000 个
  partyids?: number[]; // 公开的部门列表，最多 100 个
}
// ...
export interface ScheduleAttendee {
  userid: string;
}

export interface ScheduleReminders {
  is_remind?: 0 | 1;
  is_repeat?: 0 | 1;
  remind_before_event_secs?: number;
  remind_time_diffs?: number[];
  repeat_type?: 0 | 1 | 2 | 5 | 7;
  repeat_until?: number;
  is_custom_repeat?: 0 | 1;
  repeat_interval?: number;
  repeat_day_of_week?: number[];
  repeat_day_of_month?: number[];
  timezone?: number;
  exclude_time_list?: Array<{ start_time: number }>;
}